+++ src/index.ts
@@ -1,2 +1,11 @@
 export { Button } from "./components/button";
 export type { ButtonProps } from "./components/button";
+
+export { Input } from "./components/input";
+export type { InputProps } from "./components/input";
+
+export { Textarea } from "./components/textarea";
+export type { TextareaProps } from "./components/textarea";
+
+export { MaskedInput } from "./components/masked-input";
+export type { MaskedInputProps } from "./components/masked-input";